Explanation of merge sort:
Time Complexity: O(n logn)
Space Complexity: O(n)

Merge sort algorithm is used to sort arrays in a way that worst case time complexity is significantly reduced as compared with bubble/selection sort algorithms.

This sorting technique is based on divide and conquer approach.
A given list is broken down into two halves. The resultant lists are further broken down into two halves each. This process is carried further with the future lists until single element arrays are obtained
Now these arrays are compared (the single elements are compared), two at a time and merged as a sorted array. This process is carried until all elements are arranged in a sorted order.

Two functions are needed for this technique:
1. Sort function which divides all the elements into singles
2. Merge function which merges all the elements into a single array